+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Validate operand length in machine instructions
 - Provide bulk copybook download via Zowe Explorer
 - Support program and assembler type symbol attributes
+- Support SYSATTRA and SYSATTRP functions
 
 #### Fixed
 - Logs produced by language server in WebAssembly mode are not visible